Module3_RAG/requirements.txt
langchain==0.2.16
langchain-community==0.2.16
langchain-text-splitters==0.2.4
faiss-cpu
langchain-openai
gradio>=4.0.0
python-dotenv>=1.0.0
tiktoken>=0.7.0